pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "file-analyzer"
version = "0.2.0"
description = "A unified tool for comprehensive file analysis"
authors = [
    {name = "The File Analyzer Team"}
]
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "typer[all]>=0.9.0",
    "plumbum>=1.8.0",
    "jsonschema>=4.17.0",
    "pillow>=9.0.0",
    "pixelmatch>=0.3.0",
    "imagehash>=4.3.1",
    "scikit-image>=0.19.0",
    "python-json-logger>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.1.0",
    "syrupy>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
]

vision = [
    "mlx>=0.3.0",
    "mlx-fastvlm>=0.1.0",
    "mlx-vlm>=0.0.10",
]

[project.urls]
"Homepage" = "https://github.com/primeinc/file-analyzer"
"Bug Tracker" = "https://github.com/primeinc/file-analyzer/issues"

[project.scripts]
fa = "src.cli.main:app"

[tool.setuptools]
# Use find-packages for more reliable package discovery
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.setuptools.package-data]
"schemas" = ["**/*.json"]

[project.entry-points."fa.commands"]
analyze = "src.cli.analyze.main:app"
test = "src.cli.test.main:app"
validate = "src.cli.validate.main:app"
artifact = "src.cli.artifact.main:app"
install = "src.cli.install.main:app"
model = "src.cli.model.main:app"
benchmark = "src.cli.benchmark.main:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311"]

[tool.isort]
profile = "black"
line_length = 88